// types.ts
export interface Photo {
  id: string;
  url: string;
  isMain: boolean;
}

export type BleachingLevel = 'None' | 'Mild' | 'Medium' | 'Strong';

export interface HealthReport {
  id: string;
  date: string;
  healthPercentage: number; // 0-100
  notes: string;
  bleaching: BleachingLevel;
}

export interface GrowthReport {
  id: string;
  date: string;
  surfaceAreaM2: number; // in square meters
  volumeM3: number; // in cubic meters
}
// ...
export type Page = 'details' | 'addEditItems' | 'rules' | 'healthReports' | 'growthReports' | 'backupRestore' | 'reports' | 'speciesId' | 'archive' | 'notesToDo' | 'monitoring' | 'trees' | 'modelComparison' | 'sites' | 'anchors' | 'collectionZones' | 'branches' | 'environmental' | 'dashboard' | 'experiments' | 'treeShadeExperiment' | 'ropeOnRubbleExperiment' | 'squareRopeFrameExperiment' | 'cubeRopeFrameExperiment' | 'trends' | 'floatManagement' | 'moveItems';
export type AddEditSection = 'Sites' | 'Collection Zones' | 'Anchors' | 'Trees' | 'Branches' | 'Floats' | 'Rules' | 'Move Items';
// ...
export interface CoralBranch {
  id:string;
  fragmentId: string;
  genus: string;
  species: string;
  dateAdded: string; // ISO string format
  anchor: string;
  tree: number;
  face: 1 | 2 | 3 | 4;
  position: number;
  collectionZone: string;
  site: string;
  photos: Photo[];
  healthReports: HealthReport[];
  growthReports: GrowthReport[];
  isHeatTolerant?: boolean;
  isArchived?: boolean;
}